feat: revamp the whole architecture
