+++ package.json
@@ -11,7 +11,9 @@
 	"license": "ISC",
 	"description": "",
 	"devDependencies": {
+		"@types/cookie-parser": "^1.4.9",
 		"@types/express": "^5.0.3",
+		"@types/json2csv": "^5.0.7",
 		"@types/multer": "^2.0.0",
 		"@types/node": "^24.0.14",
 		"@types/oracledb": "^6.6.1",
