+++ src/lib/multer.ts
@@ -1,0 +1,2 @@
+import multer from "multer";
+export const upload = multer({ dest: "uploads/" });
